minor search results refactor
